requirements.txt
altair==5.5.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
boto3==1.35.69
botocore==1.35.69
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cramjam==2.9.0
cryptography==43.0.3
dnspython==2.7.0
exceptiongroup==1.2.2
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.14.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.10.1
pymongo-auth-aws==1.3.0
pymongocrypt==1.11.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-snappy==0.7.3
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
s3transfer==0.10.4
service-identity==24.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.40.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
winkerberos==0.12.0
zstandard==0.23.0
